+++ app/app.go
@@ -151,10 +151,9 @@ const (
 	// NOTE: In the SDK, the default value is 255.
 	AddrLen = 20
 
-	FlagMemIAVL  = "store.memiavl"
-	FlagAsyncWAL = "store.memiavl-async-wal"
-	// don't enable zero-copy together with inter-block cache.
-	FlagZeroCopy = "store.memiavl-zero-copy"
+	FlagMemIAVL     = "memiavl.enable"
+	FlagAsyncCommit = "memiavl.async-commit"
+	FlagZeroCopy    = "memiavl.zero-copy"
 )
 
 // this line is used by starport scaffolding # stargate/wasm/app/enabledProposals
@@ -353,7 +352,7 @@ func New(
 		// cms must be overridden before the other options, because they may use the cms,
 		// FIXME we are assuming the cms won't be overridden by the other options, but we can't be sure.
 		cms := rootmulti.NewStore(filepath.Join(homePath, "data", "memiavl.db"), logger)
-		cms.SetAsyncWAL(cast.ToBool(appOpts.Get(FlagAsyncWAL)))
+		cms.SetAsyncWAL(cast.ToBool(appOpts.Get(FlagAsyncCommit)))
 		cms.SetZeroCopy(cast.ToBool(appOpts.Get(FlagZeroCopy)))
 		baseAppOptions = append([]func(*baseapp.BaseApp){setCMS(cms)}, baseAppOptions...)
 	}